feat: 🚚 move regions constant to data-model
So we can access it for PDF Gen
