--- apps/ui/src/global/constants.ts
@@ -1,304 +1,0 @@
-/*
- * Copyright (c) 2025 The Ontario Institute for Cancer Research. All rights reserved
- *
- * This program and the accompanying materials are made available under the terms of
- * the GNU Affero General Public License v3.0. You should have received a copy of the
- * GNU Affero General Public License along with this program.
- *  If not, see <http://www.gnu.org/licenses/>.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
- * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
- * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-/**
- * List of personal titles used in the "Title" field for applicant info
- * and institutional rep info among others.
- */
-const PERSONAL_TITLES = [
-	{ en: 'Dr.', fr: '' },
-	{ en: 'Miss.', fr: '' },
-	{ en: 'Mr.', fr: '' },
-	{ en: 'Mrs.', fr: '' },
-	{ en: 'Ms.', fr: '' },
-	{ en: 'Prof.', fr: '' },
-];
-
-/**
- * @description A list of countries, territories, and geographic areas derived from The Government of Canada data reference standard.
- * Contains both English and French names, alongside ISO codes.
- *
- * Contains information licensed under The Open Government Licence – Canada.
- *
- * **Last Updated: Jan 28th, 2025**
- *
- * @license https://open.canada.ca/en/open-government-licence-canada
- * @see https://open.canada.ca/data/en/dataset/cac6fd9f-594a-4bcd-bf17-10295812d4c5
- */
-const GC_STANDARD_GEOGRAPHIC_AREAS = [
-	{ en: 'Afghanistan', fr: 'Afghanistan', iso: 'AFG' },
-	{ en: 'Åland', fr: 'Åland', iso: 'ALA' },
-	{ en: 'Albania', fr: 'Albanie', iso: 'ALB' },
-	{ en: 'Algeria', fr: 'Algérie', iso: 'DZA' },
-	{ en: 'American Samoa', fr: 'Samoa américaines', iso: 'ASM' },
-	{ en: 'Andorra', fr: 'Andorre', iso: 'AND' },
-	{ en: 'Angola', fr: 'Angola', iso: 'AGO' },
-	{ en: 'Anguilla', fr: 'Anguilla', iso: 'AIA' },
-	{ en: 'Antarctica', fr: 'Antarctique (l’)', iso: 'ATA' },
-	{ en: 'Antigua and Barbuda', fr: 'Antigua-et-Barbuda', iso: 'ATG' },
-	{ en: 'Argentina', fr: 'Argentine', iso: 'ARG' },
-	{ en: 'Armenia', fr: 'Arménie', iso: 'ARM' },
-	{ en: 'Aruba', fr: 'Aruba', iso: 'ABW' },
-	{ en: 'Australia', fr: 'Australie', iso: 'AUS' },
-	{ en: 'Austria', fr: 'Autriche', iso: 'AUT' },
-	{ en: 'Azerbaijan', fr: 'Azerbaïdjan', iso: 'AZE' },
-	{ en: 'Bahamas (The)', fr: 'Bahamas', iso: 'BHS' },
-	{ en: 'Bahrain', fr: 'Bahreïn', iso: 'BHR' },
-	{ en: 'Bangladesh', fr: 'Bangladesh', iso: 'BGD' },
-	{ en: 'Barbados', fr: 'Barbade', iso: 'BRB' },
-	{ en: 'Belarus', fr: 'Bélarus', iso: 'BLR' },
-	{ en: 'Belgium', fr: 'Belgique', iso: 'BEL' },
-	{ en: 'Belize', fr: 'Belize', iso: 'BLZ' },
-	{ en: 'Benin', fr: 'Bénin', iso: 'BEN' },
-	{ en: 'Bermuda', fr: 'Bermudes (les)', iso: 'BMU' },
-	{ en: 'Bhutan', fr: 'Bhoutan', iso: 'BTN' },
-	{ en: 'Bolivia', fr: 'Bolivie', iso: 'BOL' },
-	{ en: 'Bonaire, Sint Eustatius and Saba', fr: 'Bonaire, Saint-Eustache et Saba', iso: 'BES' },
-	{ en: 'Bosnia and Herzegovina', fr: 'Bosnie-Herzégovine', iso: 'BIH' },
-	{ en: 'Botswana', fr: 'Botswana', iso: 'BWA' },
-	{ en: 'Bouvet Island', fr: 'Bouvet (l’Île)', iso: 'BVT' },
-	{ en: 'Brazil', fr: 'Brésil', iso: 'BRA' },
-	{ en: 'British Antarctic Territory', fr: 'Antarctique, Territoire britannique de l’', iso: 'ATB' },
-	{ en: 'British Indian Ocean Territory', fr: 'Territoire britannique de l’océan Indien', iso: 'IOT' },
-	{ en: 'British Virgin Islands', fr: 'Vierges britanniques (les Îles)', iso: 'VGB' },
-	{ en: 'Brunei', fr: 'Brunéi', iso: 'BRN' },
-	{ en: 'Bulgaria', fr: 'Bulgarie', iso: 'BGR' },
-	{ en: 'Burkina Faso', fr: 'Burkina Faso', iso: 'BFA' },
-	{ en: 'Burundi', fr: 'Burundi', iso: 'BDI' },
-	{ en: 'Cabo Verde', fr: 'Cabo Verde', iso: 'CPV' },
-	{ en: 'Cambodia', fr: 'Cambodge', iso: 'KHM' },
-	{ en: 'Cameroon', fr: 'Cameroun', iso: 'CMR' },
-	{ en: 'Canada', fr: 'Canada', iso: 'CAN' },
-	{ en: 'Cayman Islands', fr: 'Caïmans (les Îles)', iso: 'CYM' },
-	{ en: 'Central African Republic', fr: 'République centrafricaine', iso: 'CAF' },
-	{ en: 'Chad', fr: 'Tchad', iso: 'TCD' },
-	{ en: 'Chile', fr: 'Chili', iso: 'CHL' },
-	{ en: 'China', fr: 'Chine', iso: 'CHN' },
-	{ en: 'Christmas Island', fr: 'Île Christmas', iso: 'CXR' },
-	{ en: 'Cocos (Keeling) Islands', fr: 'Cocos (les Îles)/ Keeling (les Îles)', iso: 'CCK' },
-	{ en: 'Colombia', fr: 'Colombie', iso: 'COL' },
-	{ en: 'Comoros', fr: 'Comores', iso: 'COM' },
-	{ en: 'Congo', fr: 'Congo', iso: 'COG' },
-	{ en: 'Cook Islands', fr: 'Îles Cook', iso: 'COK' },
-	{ en: 'Costa Rica', fr: 'Costa Rica', iso: 'CRI' },
-	{ en: 'Côte d’Ivoire', fr: 'Côte d’Ivoire', iso: 'CIV' },
-	{ en: 'Croatia', fr: 'Croatie', iso: 'HRV' },
-	{ en: 'Cuba', fr: 'Cuba', iso: 'CUB' },
-	{ en: 'Curaçao', fr: 'Curaçao', iso: 'CUW' },
-	{ en: 'Cyprus', fr: 'Chypre', iso: 'CYP' },
-	{ en: 'Czechia', fr: 'Tchéquie', iso: 'CZE' },
-	{ en: 'Democratic Republic of Congo', fr: 'République démocratique du Congo', iso: 'COD' },
-	{ en: 'Denmark', fr: 'Danemark', iso: 'DNK' },
-	{ en: 'Djibouti', fr: 'Djibouti', iso: 'DJI' },
-	{ en: 'Dominica', fr: 'Dominique', iso: 'DMA' },
-	{ en: 'Dominican Republic', fr: 'République dominicaine', iso: 'DOM' },
-	{ en: 'Ecuador', fr: 'Équateur', iso: 'ECU' },
-	{ en: 'Egypt', fr: 'Égypte', iso: 'EGY' },
-	{ en: 'El Salvador', fr: 'Salvador', iso: 'SLV' },
-	{ en: 'Equatorial Guinea', fr: 'Guinée équatoriale', iso: 'GNQ' },
-	{ en: 'Eritrea', fr: 'Érythrée', iso: 'ERI' },
-	{ en: 'Estonia', fr: 'Estonie', iso: 'EST' },
-	{ en: 'Eswatini', fr: 'Eswatini', iso: 'SWZ' },
-	{ en: 'Ethiopia', fr: 'Éthiopie', iso: 'ETH' },
-	{ en: 'Falkland Islands', fr: 'Falkland (les Îles)/Malouines (les Îles', iso: 'FLK' },
-	{ en: 'Faroe Islands', fr: 'Îles Féroé', iso: 'FRO' },
-	{ en: 'Fiji', fr: 'Fidji', iso: 'FJI' },
-	{ en: 'Finland', fr: 'Finlande', iso: 'FIN' },
-	{ en: 'France', fr: 'France', iso: 'FRA' },
-	{ en: 'French Guiana', fr: 'Guyane française', iso: 'GUF' },
-	{ en: 'French Polynesia (French Republic)', fr: 'Polynésie française (République française)', iso: 'PYF' },
-	{ en: 'French Southern and Antarctic Territories', fr: 'Terres australes et antarctiques françaises', iso: 'ATF' },
-	{ en: 'Gabon', fr: 'Gabon', iso: 'GAB' },
-	{ en: 'Gambia (The)', fr: 'Gambie', iso: 'GMB' },
-	{ en: 'Georgia', fr: 'Géorgie', iso: 'GEO' },
-	{ en: 'Germany', fr: 'Allemagne', iso: 'DEU' },
-	{ en: 'Ghana', fr: 'Ghana', iso: 'GHA' },
-	{ en: 'Gibraltar', fr: 'Gibraltar', iso: 'GIB' },
-	{ en: 'Greece', fr: 'Grèce', iso: 'GRC' },
-	{ en: 'Greenland', fr: 'Groenland (le)', iso: 'GRL' },
-	{ en: 'Grenada', fr: 'Grenade', iso: 'GRD' },
-	{ en: 'Guadeloupe', fr: 'Guadeloupe', iso: 'GLP' },
-	{ en: 'Guam', fr: 'Guam', iso: 'GUM' },
-	{ en: 'Guatemala', fr: 'Guatemala', iso: 'GTM' },
-	{ en: 'Guernsey', fr: 'Guernesey', iso: 'GGY' },
-	{ en: 'Guinea', fr: 'Guinée', iso: 'GIN' },
-	{ en: 'Guinea-Bissau', fr: 'Guinée-Bissau', iso: 'GNB' },
-	{ en: 'Guyana', fr: 'Guyana', iso: 'GUY' },
-	{ en: 'Haiti', fr: 'Haïti', iso: 'HTI' },
-	{ en: 'Heard Island and McDonald Islands', fr: 'îles Heard-et-MacDonald', iso: 'HMD' },
-	{ en: 'Honduras', fr: 'Honduras', iso: 'HND' },
-	{ en: 'Hong Kong', fr: 'Hong Kong', iso: 'HKG' },
-	{ en: 'Hungary', fr: 'Hongrie', iso: 'HUN' },
-	{ en: 'Iceland', fr: 'Islande', iso: 'ISL' },
-	{ en: 'India', fr: 'Inde', iso: 'IND' },
-	{ en: 'Indonesia', fr: 'Indonésie', iso: 'IDN' },
-	{ en: 'Iran', fr: 'Iran', iso: 'IRN' },
-	{ en: 'Iraq', fr: 'Iraq', iso: 'IRQ' },
-	{ en: 'Ireland', fr: 'Irlande', iso: 'IRL' },
-	{ en: 'Isle of Man', fr: 'Île de Man', iso: 'IMN' },
-	{ en: 'Israel', fr: 'Israël', iso: 'ISR' },
-	{ en: 'Italy', fr: 'Italie', iso: 'ITA' },
-	{ en: 'Jamaica', fr: 'Jamaïque', iso: 'JAM' },
-	{ en: 'Japan', fr: 'Japon', iso: 'JPN' },
-	{ en: 'Jersey', fr: 'Jersey', iso: 'JEY' },
-	{ en: 'Jordan', fr: 'Jordanie', iso: 'JOR' },
-	{ en: 'Kazakhstan', fr: 'Kazakhstan', iso: 'KAZ' },
-	{ en: 'Kenya', fr: 'Kenya', iso: 'KEN' },
-	{ en: 'Kiribati', fr: 'Kiribati', iso: 'KIR' },
-	{ en: 'Kosovo', fr: 'Kosovo', iso: 'XKX' },
-	{ en: 'Kuwait', fr: 'Koweït', iso: 'KWT' },
-	{ en: 'Kyrgyzstan', fr: 'Kirghizistan', iso: 'KGZ' },
-	{ en: 'Laos', fr: 'Laos', iso: 'LAO' },
-	{ en: 'Latvia', fr: 'Lettonie', iso: 'LVA' },
-	{ en: 'Lebanon', fr: 'Liban', iso: 'LBN' },
-	{ en: 'Lesotho', fr: 'Lesotho', iso: 'LSO' },
-	{ en: 'Liberia', fr: 'Libéria', iso: 'LBR' },
-	{ en: 'Libya', fr: 'Libye', iso: 'LBY' },
-	{ en: 'Liechtenstein', fr: 'Liechtenstein', iso: 'LIE' },
-	{ en: 'Lithuania', fr: 'Lituanie', iso: 'LTU' },
-	{ en: 'Luxembourg', fr: 'Luxembourg', iso: 'LUX' },
-	{ en: 'Macao', fr: 'Macao', iso: 'MAC' },
-	{ en: 'Madagascar', fr: 'Madagascar', iso: 'MDG' },
-	{ en: 'Malawi', fr: 'Malawi', iso: 'MWI' },
-	{ en: 'Malaysia', fr: 'Malaisie', iso: 'MYS' },
-	{ en: 'Maldives', fr: 'Maldives', iso: 'MDV' },
-	{ en: 'Mali', fr: 'Mali', iso: 'MLI' },
-	{ en: 'Malta', fr: 'Malte', iso: 'MLT' },
-	{ en: 'Marshall Islands', fr: 'Îles Marshall', iso: 'MHL' },
-	{ en: 'Martinique', fr: 'Martinique', iso: 'MTQ' },
-	{ en: 'Mauritania', fr: 'Mauritanie', iso: 'MRT' },
-	{ en: 'Mauritius', fr: 'Maurice', iso: 'MUS' },
-	{ en: 'Mayotte', fr: 'Mayotte', iso: 'MYT' },
-	{ en: 'Mexico', fr: 'Mexique', iso: 'MEX' },
-	{ en: 'Micronesia', fr: 'Micronésie', iso: 'FSM' },
-	{ en: 'Moldova', fr: 'la Moldova', iso: 'MDA' },
-	{ en: 'Monaco', fr: 'Monaco', iso: 'MCO' },
-	{ en: 'Mongolia', fr: 'Mongolie', iso: 'MNG' },
-	{ en: 'Montenegro', fr: 'Monténégro', iso: 'MNE' },
-	{ en: 'Montserrat', fr: 'Montserrat', iso: 'MSR' },
-	{ en: 'Morocco', fr: 'Maroc', iso: 'MAR' },
-	{ en: 'Mozambique', fr: 'Mozambique', iso: 'MOZ' },
-	{ en: 'Myanmar', fr: 'Myanmar', iso: 'MMR' },
-	{ en: 'Namibia', fr: 'Namibie', iso: 'NAM' },
-	{ en: 'Nauru', fr: 'Nauru', iso: 'NRU' },
-	{ en: 'Nepal', fr: 'Népal', iso: 'NPL' },
-	{ en: 'Netherlands', fr: 'Pays-Bas', iso: 'NLD' },
-	{ en: 'New Caledonia', fr: 'Nouvelle-Calédonie', iso: 'NCL' },
-	{ en: 'New Zealand', fr: 'Nouvelle-Zélande', iso: 'NZL' },
-	{ en: 'Nicaragua', fr: 'Nicaragua', iso: 'NIC' },
-	{ en: 'Niger', fr: 'Niger', iso: 'NER' },
-	{ en: 'Nigeria', fr: 'Nigéria', iso: 'NGA' },
-	{ en: 'Niue', fr: 'Niue', iso: 'NIU' },
-	{ en: 'Norfolk Island', fr: 'Norfolk (l’Île)', iso: 'NFK' },
-	{ en: 'North Korea', fr: 'Corée du Nord', iso: 'PRK' },
-	{ en: 'North Macedonia', fr: 'Macédoine du Nord', iso: 'MKD' },
-	{ en: 'Northern Mariana Islands', fr: 'Mariannes du Nord (les Îles)', iso: 'MNP' },
-	{ en: 'Norway', fr: 'Norvège', iso: 'NOR' },
-	{ en: 'Oman', fr: 'Oman', iso: 'OMN' },
-	{ en: 'Pakistan', fr: 'Pakistan', iso: 'PAK' },
-	{ en: 'Palau', fr: 'Palaos', iso: 'PLW' },
-	{ en: 'Panama', fr: 'Panama', iso: 'PAN' },
-	{ en: 'Papua New Guinea', fr: 'Papouasie-Nouvelle-Guinée', iso: 'PNG' },
-	{ en: 'Paraguay', fr: 'Paraguay', iso: 'PRY' },
-	{ en: 'Peru', fr: 'Pérou', iso: 'PER' },
-	{ en: 'Philippines', fr: 'Philippines', iso: 'PHL' },
-	{ en: 'Pitcairn', fr: 'Pitcairn', iso: 'PCN' },
-	{ en: 'Poland', fr: 'Pologne', iso: 'POL' },
-	{ en: 'Portugal', fr: 'Portugal', iso: 'PRT' },
-	{ en: 'Puerto Rico', fr: 'Porto Rico', iso: 'PRI' },
-	{ en: 'Qatar', fr: 'Qatar', iso: 'QAT' },
-	{ en: 'Queen Maud Land', fr: 'Terre de la Reine Maud', iso: 'ATN' },
-	{ en: 'Republic of Korea', fr: 'République de Corée', iso: 'KOR' },
-	{ en: 'Reunion', fr: 'Réunion (La)', iso: 'REU' },
-	{ en: 'Romania', fr: 'Roumanie', iso: 'ROU' },
-	{ en: 'Russia', fr: 'Russie', iso: 'RUS' },
-	{ en: 'Rwanda', fr: 'Rwanda', iso: 'RWA' },
-	{ en: 'Saint Barthélemy (French Republic)', fr: 'Saint-Barthélemy (République française)', iso: 'BLM' },
-	{ en: 'Saint Helena Ascension and Tristan da Cunha', fr: 'Sainte-Hélène Ascension et Tristan da Cunha', iso: 'SHN' },
-	{ en: 'Saint Kitts and Nevis', fr: 'Saint-Kitts-et-Nevis', iso: 'KNA' },
-	{ en: 'Saint Lucia', fr: 'Sainte-Lucie', iso: 'LCA' },
-	{ en: 'Saint Martin (French Republic)', fr: 'Saint-Martin (République française)', iso: 'MAF' },
-	{
-		en: 'Saint Pierre and Miquelon (French Republic)',
-		fr: 'Saint-Pierre-et-Miquelon (République française)',
-		iso: 'SPM',
-	},
-	{ en: 'Saint Vincent and the Grenadines', fr: 'Saint-Vincent-et-les Grenadines', iso: 'VCT' },
-	{ en: 'Samoa', fr: 'Samoa', iso: 'WSM' },
-	{ en: 'San Marino', fr: 'Saint-Marin', iso: 'SMR' },
-	{ en: 'São Tomé and Príncipe', fr: 'Sao Tomé-et-Principe', iso: 'STP' },
-	{ en: 'Sark', fr: 'Sercq', iso: 'CRQ' },
-	{ en: 'Saudi Arabia', fr: 'Arabie saoudite', iso: 'SAU' },
-	{ en: 'Senegal', fr: 'Sénégal', iso: 'SEN' },
-	{ en: 'Serbia', fr: 'Serbie', iso: 'SRB' },
-	{ en: 'Seychelles', fr: 'Seychelles', iso: 'SYC' },
-	{ en: 'Sierra Leone', fr: 'Sierra Leone', iso: 'SLE' },
-	{ en: 'Singapore', fr: 'Singapour', iso: 'SGP' },
-	{ en: 'Sint Maarten', fr: 'Saint-Martin', iso: 'SXM' },
-	{ en: 'Slovakia', fr: 'Slovaquie', iso: 'SVK' },
-	{ en: 'Slovenia', fr: 'Slovénie', iso: 'SVN' },
-	{ en: 'Solomon Islands', fr: 'Îles Salomon', iso: 'SLB' },
-	{ en: 'Somalia', fr: 'Somalie', iso: 'SOM' },
-	{ en: 'South Africa', fr: 'Afrique du Sud', iso: 'ZAF' },
-	{ en: 'South Georgia and the South Sandwich Islands', fr: 'Géorgie du Sud-et-les Îles Sandwich du Sud', iso: 'SGS' },
-	{ en: 'South Sudan', fr: 'Soudan du Sud', iso: 'SSD' },
-	{ en: 'Spain', fr: 'Espagne', iso: 'ESP' },
-	{ en: 'Sri Lanka', fr: 'Sri Lanka', iso: 'LKA' },
-	{ en: 'Sudan', fr: 'Soudan', iso: 'SDN' },
-	{ en: 'Suriname', fr: 'Suriname', iso: 'SUR' },
-	{ en: 'Svalbard and Jan Mayen Islands', fr: 'Îles Svalbard-et-Jan Mayen', iso: 'SJM' },
-	{ en: 'Sweden', fr: 'Suède', iso: 'SWE' },
-	{ en: 'Switzerland', fr: 'Suisse', iso: 'CHE' },
-	{ en: 'Syria', fr: 'Syrie', iso: 'SYR' },
-	{ en: 'Taiwan', fr: 'Taïwan', iso: 'TWN' },
-	{ en: 'Tajikistan', fr: 'Tadjikistan', iso: 'TJK' },
-	{ en: 'Tanzania', fr: 'Tanzanie', iso: 'TZA' },
-	{ en: 'Thailand', fr: 'Thaïlande', iso: 'THA' },
-	{ en: 'Timor-Leste', fr: 'Timor-Leste', iso: 'TLS' },
-	{ en: 'Togo', fr: 'Togo', iso: 'TGO' },
-	{ en: 'Tokelau', fr: 'Tokelau', iso: 'TKL' },
-	{ en: 'Tonga', fr: 'Tonga', iso: 'TON' },
-	{ en: 'Trinidad and Tobago', fr: 'Trinité-et-Tobago', iso: 'TTO' },
-	{ en: 'Tunisia', fr: 'Tunisie', iso: 'TUN' },
-	{ en: 'Türkiye', fr: 'Türkiye', iso: 'TUR' },
-	{ en: 'Turkmenistan', fr: 'Turkménistan', iso: 'TKM' },
-	{ en: 'Turks and Caicos Islands', fr: 'Turks-et-Caïcos (les Îles)', iso: 'TCA' },
-	{ en: 'Tuvalu', fr: 'Tuvalu', iso: 'TUV' },
-	{ en: 'Uganda', fr: 'Ouganda', iso: 'UGA' },
-	{ en: 'Ukraine', fr: 'Ukraine', iso: 'UKR' },
-	{ en: 'United Arab Emirates', fr: 'Émirats arabes unis', iso: 'ARE' },
-	{ en: 'United Kingdom', fr: 'Royaume-Uni', iso: 'GBR' },
-	{ en: 'United States', fr: 'États-Unis', iso: 'USA' },
-	{ en: 'United States Minor Outlying Islands', fr: 'Îles mineures éloignées des États-Unis (les)', iso: 'UMI' },
-	{ en: 'Uruguay', fr: 'Uruguay', iso: 'URY' },
-	{ en: 'US Virgin Islands', fr: 'Vierges des États-Unis (les Îles)', iso: 'VIR' },
-	{ en: 'Uzbekistan', fr: 'Ouzbékistan', iso: 'UZB' },
-	{ en: 'Vanuatu', fr: 'Vanuatu', iso: 'VUT' },
-	{ en: 'Vatican City State', fr: 'État de la Cité du Vatican', iso: 'VAT' },
-	{ en: 'Venezuela', fr: 'Venezuela', iso: 'VEN' },
-	{ en: 'Vietnam', fr: 'Vietnam', iso: 'VNM' },
-	{ en: 'Wallis and Futuna (French Republic)', fr: 'Wallis-et-Futuna (République française)', iso: 'WLF' },
-	{ en: 'West Bank and Gaza', fr: 'Cisjordanie et Gaza', iso: 'PSE' },
-	{ en: 'Western Sahara', fr: 'Sahara occidental (le)', iso: 'ESH' },
-	{ en: 'Yemen', fr: 'Yémen', iso: 'YEM' },
-	{ en: 'Zambia', fr: 'Zambie', iso: 'ZMB' },
-	{ en: 'Zimbabwe', fr: 'Zimbabwe', iso: 'ZWE' },
-];
-
-export { GC_STANDARD_GEOGRAPHIC_AREAS, PERSONAL_TITLES };
