+++ packages/data-model/src/constants/index.ts
@@ -17,4 +17,4 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export * from './personalTitles';
+export * from './countriesAndRegions.js';
