+++ packages/data-model/src/main.ts
@@ -17,6 +17,6 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+export * from './constants/index.js';
 export * from './types.js';
-
 export { isApplicationStateValue } from './utils.js';
